Add support for hussh and ansible-pylibssh clients.
